+++ resources/leiningen/new/luminus/keeframe/src/cljs/view.cljs
@@ -38,8 +38,6 @@
 
 (defn home-page []
   [:div.container
-   [:div.row>div.col-sm-12
-    [:h2.alert.alert-info "Tip: try pressing CTRL+H to open re-frame tracing menu"]]
    (when-let [docs @(rf/subscribe [:docs])]
      [:div.row>div.col-sm-12
       [:div {:dangerouslySetInnerHTML