+++ src/leiningen/new/db.clj
@@ -11,8 +11,8 @@
     (some #{"+sqlite"} features) :sqlite))
 
 (defn db-dependencies [options]
-  (into [['luminus-migrations "0.6.0"]
-         ['conman "0.8.2"]]
+  (into [['luminus-migrations "0.6.1"]
+         ['conman "0.8.3"]]
         ({:postgres [['org.postgresql/postgresql "42.2.5"]]
           :mysql    [['mysql/mysql-connector-java "8.0.12"]
                      ['com.google.protobuf/protobuf-java "3.6.1"]]